Improve WebUI tests
* centralize utility functions
* use IDs wherever possible
* remove console.log
* remove unnecessary page reload
