+++ src/webapp/components/Header.js
@@ -226,6 +226,7 @@ export default function Header(props) {
         <IconButton
           color="inherit"
           className="help-button"
+          id="help-button"
           onClick={() => {
             setDialogOpen(true);
           }}
